1. Refactors files.
